pose-vision/requirements.txt
viam-sdk
pydantic
joblib
scikit-learn
